// src/models/types/MonsterKey.ts
enum MonsterKey {
  ABADDON = "ABADDON",
  ABYSS_WORM = "ABYSS_WORM",
  ACHELOUS = "ACHELOUS",
  ADAMANTOISE = "ADAMANTOISE",
  AEROUGE = "AEROGUE",
  AHRIMAN = "AHRIMAN",
  ALCYONE = "ALCYONE",
  ANACONDAUR = "ANACONDAUR",
  BANDERSNATCH = "BANDERSNATCH",
  BARBATOS = "BARBATOS",
  BASHURA = "BASHURA",
  BASILISK = "BASILISK",
  BEHEMOTH = "BEHEMOTH",
  BEHEMOTH_KING = "BEHEMOTH_KING",
  BITE_BUG = "BITE_BUG",
  BLACK_ELEMENT = "BLACK_ELEMENT",
  BLUE_ELEMENT = "BLUE_ELEMENT",
  BOMB = "BOMB",
  BOMB_KING = "BOMB_KING",
  BUER = "BUER",
  BUNYIP = "BUNYIP",
  CACTUAR = "CACTUAR",
  CACTUAR_KING = "CACTUAR_KING",
  CATOBLEPAS = "CATOBEPLAS",
  CHIMERA = "CHIMERA",
  CHIMERA_BRAIN = "CHIMERA_BRAIN",
  CHIMERAGEIST = "CHIMERAGEIST",
  COEURL = "COEURL",
  COEURLREGINA = "COEURLREGINA",
  CONDOR = "CONDOR",
  DARK_ELEMENT = "DARK_ELEMENT",
  DARK_FLAN = "DARK_FLAN",
  DEMONOLITH = "DEMONOLITH",
  DINGO = "DINGO",
  DINONIX = "DINONIX",
  DON_TONBERRY = "DON_TONBERRY",
  DUAL_HORN = "DUAL_HORN",
  EPAAJ = "EPAAJ",
  ESPADA = "ESPADA",
  EVIL_EYE = "EVIL_EYE",
  EXORAY = "EXORAY",
  FAFNIR = "FAFNIR",
  FENRIR = "FENRIR",
  FLAME_FLAN = "FLAME_FLAN",
  FLOATING_DEATH = "FLOATING_DEATH",
  FLOATING_EYE = "FLOATING_EYE",
  FUNGUAR = "FUNGUAR",
  GANDAREWA = "GANDAREWA",
  GARM = "GARM",
  GARUDA = "GARUDA",
  GEMINI_CLUB = "GEMINI_CLUB",
  GEMINI_SWORD = "GEMINI_SWORD",
  GHOST = "GHOST",
  GOLD_ELEMENT = "GOLD_ELEMENT",
  GRAT = "GRAT",
  GREAT_MALBORO = "GREAT_MALBORO",
  GRENADE = "GRENADE",
  GRENDEL = "GRENDEL",
  HALMA = "HALMA",
  HORNET = "HORNET",
  ICE_FLAN = "ICE_FLAN",
  IMP = "IMP",
  IRONCLAD = "IRONCLAD",
  IPIRIA = "IPIRIA",
  IRON_GIANT = "IRON_GIANT",
  IGUION = "IGUION",
  JORMUNGAND = "JORMUNGAND",
  JUGGERNAUT = "JUGGERNAUT",
  JUMBO_FLAN = "JUMBO_FLAN",
  KILLER_BEE = "KILLER_BEE",
  KOTTOS = "KOTTOS",
  KUSARIQQU = "KUSARIQQU",
  LAMASHTU = "LAMASHTU",
  LARVA = "LARVA",
  ONE_EYE = "ONE_EYE",
  MACHEA = "MACHEA",
  MAELSPIKE = "MAELSPIKE",
  MAFDET = "MAFDET",
  MALBORO = "MALBORO",
  MALBORO_MENACE = "MALBORO_MENACE",
  MANDRAGORA = "MANDRAGORA",
  MASTER_COEURL = "MASTER_COEURL",
  MASTER_TONBERRY = "MASTER_TONBERRY",
  MELUSINE = "MELUSINE",
  MI_IHEN_FANG = "MI_IHEN_FANG",
  MURUSSU = "MURUSSU",
  MUSHUSSU = "MUSHUSSU",
  NAGA_ELEMENTAL = "NAGA_ELEMENTAL",
  NEBIROS = "NEBIROS",
  NIDHOGG = "NIDHOGG",
  OCHU = "OCHU",
  OGRE = "OGRE",
  ORNITHOLESTES = "ORNITHOLESTES",
  PTERYX = "PTERYX",
  PUROBOROS = "PUROBOROS",
  QACTUAR = "QACTUAR",
  RAGORA = "RAGORA",
  RALDO = "RALDO",
  RAPTOR = "RAPTOR",
  RED_ELEMENT = "RED_ELEMENT",
  SAND_WOLF = "SAND_WOLF",
  SAND_WORM = "SAND_WORM",
  SHRED = "SHRED",
  SIMURGH = "SIMURGH",
  SKOLL = "SKOLL",
  SLEEP_SPROUT = "SLEEP_SPROUT",
  SNOW_FLAN = "SNOW_FLAN",
  SNOW_WOLF = "SNOW_WOLF",
  SPLASHER = "SPLASHER",
  SPIRIT = "SPIRIT",
  STRATOAVIS = "STRATOAVIS",
  TANKET = "TANKET",
  THORN = "THORN",
  THUNDER_FLAN = "THUNDER_FLAN",
  TONBERRY = "TONBERRY",
  VALAHA = "VALAHA",
  VARUNA = "VARUNA",
  VIDATU = "VIDATU",
  VORBAN = "VORBAN",
  VOUIVRE = "VOUIVRE",
  WASP = "WASP",
  WATER_FLAN = "WATER_FLAN",
  WHITE_ELEMENT = "WHITE_ELEMENT",
  WRAITH = "WRAITH",
  XIPHOS = "XIPHOS",
  YELLOW_ELEMENT = "YELLOW_ELEMENT",
  YOWIE = "YOWIE",
  ZAURUS = "ZAURUS",
  ZU = "ZU",
}

export { MonsterKey };