// src/models/types/monsterKey.ts
enum MonsterKey {
  ACHELOUS = "ACHELOUS",
  ADAMANTOISE = "ADAMANTOISE",
  AEROGUE = "AEROGUE",
  AHRIMAN = "AHRIMAN",
  ALCYONE = "ALCYONE",
  ANACONDAUR = "ANACONDAUR",
  BANDERSNATCH = "BANDERSNATCH",
  BARBATOS = "BARBATOS",
  BASHURA = "BASHURA",
  BASILISK = "BASILISK",
  BEHEMOTH = "BEHEMOTH",
  BEHEMOTH_KING = "BEHEMOTH_KING",
  BITE_BUG = "BITE_BUG",
  BLACK_ELEMENT = "BLACK_ELEMENT",
  BLUE_ELEMENT = "BLUE_ELEMENT",
  BOMB = "BOMB",
  BUER = "BUER",
  BUNYIP = "BUNYIP",
  CACTUAR = "CACTUAR",
  CHIMERA = "CHIMERA",
  CHIMERA_BRAIN = "CHIMERA_BRAIN",
  COEURL = "COEURL",
  CONDOR = "CONDOR",
  DARK_ELEMENT = "DARK_ELEMENT",
  DARK_FLAN = "DARK_FLAN",
  DEMONOLITH = "DEMONOLITH",
  DINGO = "DINGO",
  DINONIX = "DINONIX",
  DUAL_HORN = "DUAL_HORN",
  EPAAJ = "EPAAJ",
  EVIL_EYE = "EVIL_EYE",
  EXORAY = "EXORAY",
  FLAME_FLAN = "FLAME_FLAN",
  FLOATING_DEATH = "FLOATING_DEATH",
  FLOATING_EYE = "FLOATING_EYE",
  FUNGUAR = "FUNGUAR",
  GANDAREWA = "GANDAREWA",
  GARM = "GARM",
  GARUDA = "GARUDA",
  GEMINI_CLUB = "GEMINI_CLUB",
  GEMINI_SWORD = "GEMINI_SWORD",
  GHOST = "GHOST",
  GOLD_ELEMENT = "GOLD_ELEMENT",
  GRAT = "GRAT",
  GREAT_MALBORO = "GREAT_MALBORO",
  GRENADE = "GRENADE",
  GRENDEL = "GRENDEL",
  HALMA = "HALMA",
  ICE_FLAN = "ICE_FLAN",
  IMP = "IMP",
  IPIRIA = "IPIRIA",
  IRON_GIANT = "IRON_GIANT",
  IGUION = "IGUION",
  KILLER_BEE = "KILLER_BEE",
  KUSARIQQU = "KUSARIQQU",
  LAMASHTU = "LAMASHTU",
  LARVA = "LARVA",
  MACHEA = "MACHEA",
  MAELSPIKE = "MAELSPIKE",
  MAFDET = "MAFDET",
  MALBORO = "MALBORO",
  MANDRAGORA = "MANDRAGORA",
  MASTER_COEURL = "MASTER_COEURL",
  MASTER_TONBERRY = "MASTER_TONBERRY",
  MELUSINE = "MELUSINE",
  MI_IHEN_FANG = "MI_IHEN_FANG",
  MURUSSU = "MURUSSU",
  MUSHUSSU = "MUSHUSSU",
  NEBIROS = "NEBIROS",
  NIDHOGG = "NIDHOGG",
  OCHU = "OCHU",
  OGRE = "OGRE",
  PUROBOROS = "PUROBOROS",
  QACTUAR = "QACTUAR",
  RAGORA = "RAGORA",
  RALDO = "RALDO",
  RAPTOR = "RAPTOR",
  RED_ELEMENT = "RED_ELEMENT",
  SAND_WOLF = "SAND_WOLF",
  SAND_WORM = "SAND_WORM",
  SHRED = "SHRED",
  SIMURGH = "SIMURGH",
  SKOLL = "SKOLL",
  SNOW_FLAN = "SNOW_FLAN",
  SNOW_WOLF = "SNOW_WOLF",
  SPLASHER = "SPLASHER",
  SPIRIT = "SPIRIT",
  THORN = "THORN",
  THUNDER_FLAN = "THUNDER_FLAN",
  TONBERRY = "TONBERRY",
  VALAHA = "VALAHA",
  VARUNA = "VARUNA",
  VOUIVRE = "VOUIVRE",
  WASP = "WASP",
  WATER_FLAN = "WATER_FLAN",
  WHITE_ELEMENT = "WHITE_ELEMENT",
  WRAITH = "WRAITH",
  XIPHOS = "XIPHOS",
  YELLOW_ELEMENT = "YELLOW_ELEMENT",
  YOWIE = "YOWIE",
  ZAURUS = "ZAURUS",
  ZU = "ZU",
}

export { MonsterKey };